Add files via upload
I’ve created a detailed, day-wise lesson plan for those interested in studying Data Science. As an educator, I developed this plan based on my experience, and it was also used during my freelancing work. I hope this resource proves helpful to you.

This document is offered free of charge. If you choose to share it with others or post it publicly, please ensure the creator receives appropriate credit for the effort and planning that went into its creation.